Parts of GeneralCache Unit Tests is Ok The GeneralCache can not pass basic tests such as checking the functionality of allocating and writing-back. The SimpleMem also pass basic random access testing.
More complicated and robust tests are WIP.

+++ src/pipeline/mem/general_cache/replacement_policy.rs
@@ -54,7 +54,7 @@ pub mod random {
                 }
             }
             // generate a ranged usize random number
-            let evict_way = self.rng.random_range(0..self.valid.len()) as usize;
+            let evict_way = self.rng.random_range(0..self.valid.len());
             self.valid[evict_way] = false;
             (true, evict_way)
         }
